+++ agent_build_refactored/managed_packages/steps/build_python.sh
@@ -59,8 +59,24 @@ pushd build
 
 make -j "$(nproc)"
 make test
-make DESTDIR="${STEP_OUTPUT_PATH}" install
+make install
 
 popd
 popd
 popd
+
+# Install version of pip that we need.
+export LD_LIBRARY_PATH="${INSTALL_PREFIX}/lib:${LD_LIBRARY_PATH}"
+"${INSTALL_PREFIX}/bin/python3" -m pip install "pip==${PIP_VERSION}"
+
+# Remove some of the files to reduce size of the interpreter
+PYTHON_LIBS_PATH="${INSTALL_PREFIX}/lib/python${PYTHON_SHORT_VERSION}"
+
+find "${PYTHON_LIBS_PATH}" -name "__pycache__" -type d -prune -exec rm -r {} \;
+
+rm -r "${PYTHON_LIBS_PATH}/test"
+rm -r "${PYTHON_LIBS_PATH}"/config-"${PYTHON_SHORT_VERSION}"-*-linux-gnu/
+rm -r "${PYTHON_LIBS_PATH}/lib2to3"
+
+mkdir -p "${STEP_OUTPUT_PATH}${INSTALL_PREFIX}"
+cp -a "${INSTALL_PREFIX}/." "${STEP_OUTPUT_PATH}${INSTALL_PREFIX}"