+++ tests/end_to_end_tests/managed_packages_tests/test_remotely.py
@@ -21,6 +21,7 @@ import tarfile
 
 import pytest
 
+from agent_build_refactored.tools.constants import Architecture
 from tests.end_to_end_tests.run_in_remote_machine import run_test_remotely
 from tests.end_to_end_tests.run_in_remote_machine.portable_pytest_runner import (
     PORTABLE_PYTEST_RUNNER_BUILDERS,
@@ -41,9 +42,14 @@ def test_remotely(
     scalyr_server,
     test_session_suffix,
     tmp_path,
+    use_aio_package,
 ):
 
-    arch = package_builder.DEPENDENCY_PACKAGES_ARCHITECTURE
+    if use_aio_package:
+        arch = package_builder.DEPENDENCY_PACKAGES_ARCHITECTURE
+    else:
+        arch = Architecture.X86_64
+
     pytest_runner_builder_cls = PORTABLE_PYTEST_RUNNER_BUILDERS[arch]
     pytest_runner_builder = pytest_runner_builder_cls()
     pytest_runner_builder.build()
@@ -78,7 +84,7 @@ def test_remotely(
                 "--test-session-suffix",
                 test_session_suffix,
             ],
-            architecture=package_builder.DEPENDENCY_PACKAGES_ARCHITECTURE,
+            architecture=arch,
             pytest_runner_path=pytest_runner_builder.result_runner_path,
             file_mappings={str(packages_archive_path): "/tmp/packages.tar"},
         )