+++ RELEASE_NOTES.md
@@ -9,7 +9,7 @@ incompatible changes have been made in that release.
 For a list of all the changes in a particular release, please refer to the changelog file -
 https://github.com/scalyr/scalyr-agent-2/blob/master/CHANGELOG.md.
 
-## 2.2.1 "TBD"
+## 2.2.1 "Frosty"
 
 * This release significantly changes the Agent's `deb` and `rpm` packages. Agent now is shipped with its own, independent 
   version of the Python interpreter and does not rely on system Python. The upgrade process to the new packages has to be